Extract casbin specifics to separate file
