+++ __tests__/casbin_authorizer.test.ts
@@ -1,6 +1,7 @@
 import "jest"
 import { UserAuthInfo } from "../src/auth/authn";
-import { CasbinAuthorizer, PermissionDeniedError, Action, ReadAction, CreateAction, UpdateAction, DeleteAction } from "../src/auth/authz";
+import { PermissionDeniedError, Action, ReadAction, CreateAction, UpdateAction, DeleteAction } from "../src/auth/authz";
+import { CasbinAuthorizer } from "../src/auth/casbin";
 import { resolve } from "path";
